+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Spludo is copyright 2009-2010 by DracoBlue <http://dracoblue.net>
 * added "spludo-gen controller" (wizard to create a new controller)
 * views can now be in sub folders
 * added partial + partials in .ejs
+* .ejs-Views recompile the view as soon as it is changed
 
 ## 1.0.0 (2010/09/20)